refactor: merge units types with tokens util
