+++ src/features/swap/SwapForm.tsx
@@ -8,8 +8,7 @@ import { CostsSummary } from 'src/features/swap/CostsSummary';
 import { ReceiveSummary } from 'src/features/swap/ReceiveSummary';
 import { TokenCard } from 'src/features/swap/TokenCard';
 import { useExchangeContext } from 'src/providers/ExchangeProvider';
-import { CeloWei, StCeloWei } from 'src/types/units';
-import { fromWeiRounded } from 'src/utils/tokens';
+import { CeloWei, fromWeiRounded, StCeloWei } from 'src/utils/tokens';
 import { BalanceTools } from './BalanceTools';
 import { SubmitButton } from './SubmitButton';
 import { StakeToken } from './types';
