+++ src/features/swap/useFormValidator.ts
@@ -1,7 +1,6 @@
 import { useCallback } from 'react';
 import { StakeToken } from 'src/features/swap/types';
-import { Celo, CeloWei, StCeloWei } from 'src/types/units';
-import { toCeloWei } from 'src/utils/tokens';
+import { Celo, CeloWei, StCeloWei, toCeloWei } from 'src/utils/tokens';
 
 export function useFormValidator(balance: CeloWei | StCeloWei, token: StakeToken) {
   return useCallback(
