+++ src/hooks/useTokenBalances.ts
@@ -1,7 +1,6 @@
 import { useCallback, useEffect, useState } from 'react';
 import { useContracts } from 'src/hooks/useContracts';
-import { Celo, CeloWei } from 'src/types/units';
-import { fromCeloWei } from 'src/utils/tokens';
+import { Celo, CeloWei, fromCeloWei } from 'src/utils/tokens';
 
 export const useTokenBalances = () => {
   const { accountContract } = useContracts();
